// express/backend/src/iobroker.d.ts
export interface LatestAdapter {
	meta: string;
	icon: string;
	type: string;
	stars: number;
	weekDownloads: number;
	stat: number;
	name: string;
	version: string;
	news: TranslatedText;
	title: string;
	titleLang: TranslatedText;
	desc: TranslatedText;
	authors: string[];
	keywords: string[];
	license: string;
	platform: string;
	main: string;
	enabled: boolean;
	extIcon: string;
	readme: string;
	loglevel: string;
	mode: string;
	compact: boolean;
	connectionType: string;
	dataSource: string;
	materialize: boolean;
	dependencies: Record<string, string>[];
	published: Date;
	versionDate: Date;
	stable: string;
}

export type LatestAdapters = Record<string, LatestAdapter>;

export type AdapterName = string;

export type Version = string;

export interface Statistics {
	total: number;
	adapters: Record<AdapterName, number>;
	multihosts: number;
	platforms: Record<string, number>;
	languages: Record<string, number>;
	versions: Record<AdapterName, Record<Version, number>>;
	countries: Record<string, number>;
	counts: Record<string, number>;
	nodes: Record<string, number>;
	date: string;
}